move processEdge into util
