+++ src/graph/util.ts
@@ -259,3 +259,36 @@ export const removeAll = <T>(
   array.length -= removed.length;
   return removed;
 };
+
+const EDGE_REGEX = /(.+)->(.+)/;
+
+export interface ProcessedEdge {
+  edge: string;
+  d: string;
+}
+export const processEdge = (
+  container: ParentNode,
+  edge: `${string}->${string}`,
+  offset: Vector2
+): ProcessedEdge | null => {
+  const info = edge.match(EDGE_REGEX);
+  if (!info) {
+    throw new Error(
+      `Invalid edge '${edge}', the format should be \`\${source}.\${output}->\${destination}.\${input}`
+    );
+  }
+  const [src, dst] = info.slice(1);
+  const from = $`.output > div[data-id='${src}']`(container);
+  const to = $`.input > div[data-id='${dst}']`(container);
+  if (!from || !to) {
+    const [which, io] = from ? ["destination", "input"] : ["source", "input"];
+    throw new Error(`Invalid edge '${edge}', ${which} node does not exist or has no such ${io}`);
+  }
+  if (from.parentElement!.dataset.type === "group" || to.parentElement!.dataset.type === "group") {
+    const which = from.parentElement!.dataset.type === "group" ? "source" : "destination";
+    throw new Error(`Invalid edge '${edge}', ${which} node is inside a group`);
+  }
+  const srcMid = v2.rectMid(from.getBoundingClientRect()).subtract(offset);
+  const dstMid = v2.rectMid(to.getBoundingClientRect()).subtract(offset);
+  return { edge, d: path.bezier(srcMid, dstMid) };
+};
